エンコーダのパルス設定をfloatにした

+++ libraries/inc/Localization.hpp
@@ -35,7 +35,7 @@ public:
     virtual float GetYaw();
     void SetEncoderModef(float x_of_y, float y_of_x); //エンコーダの機体中心からのズレを記述
     void SetDiameter(float d);
-    void SetEncPulse(short p);
+    void SetEncPulse(float p);
 };
 
 #endif
